The offending segfault was due to game_memory in SDL_map_gen.c being memset with 'sizeof(memset)'. I don't know why I thought this would work... The main game loop in map_gen.c was dependant on it's values being initialised to zero, so of course it caused problems when only the first 8 bytes were being initialised like this (and also probably explains why the problem was so intermittant -- sometimes the program ran fine).
Testing the program showed that 'player_1.has_moved' was returning true, and maybe the 'initialised' variable was being intialised to true, causing problems (I didn't test it, but it seems likely). A useful lesson, because it now seems to run without any bugs.

I also added a feature to allow quick moving between rooms, so you can stand on an entrance or exit and pop back in to the previous room without needing to step outside of the entrance/exit square and then back in.