Okay, so render_game is now rendering -only- the parts of the level that fit on screen. The previous versions always started rendering the top/left sides of the map, and would have gone on to render the rest of the width on to the next line(subsequently blocked out by +row at the end of each vertical loop), and the heigt would have tried to write to as much memory as it possibly could, which is obviously bad, hence the seg faults at large sizes previously.
Now the code seems to be able to render map sizes in the thousands, but I am not a fan of the structure of the code, and it is clearly very inefficient the larger you go with map size. It'll never be necessary to render maps as large as thousands of tiles wide, but I would still like to clean up the code, making it more legible (there's no need for 'padding_width/height' variables, for example, and calculating the area to render before looping through each tile.