Added new render transitions, so that levels fade in and out when you step on or off entrance and exit tiles. If the level hasn't completed fading in when you enter the level, it will continue to do so at the set rate. There is the option of making the level 'snap' to fullly rendered when you enter the room, which is simple to implement and may be worth taking up at some point, but for the time being I am going to stick with this option as it gives an an effect of the player's eyes 'adjusting to the new level', so to speak.
When you increase the transition time to something like 100 (ms),
you can see the second to last level (the former current_level's
prev_level) snap to zero visibility, as that is how the program
currently handles the render transitions of former prev_levels.
It makes them zero to ensure that when travelling backwards, you
don't see a ghost of the old transition counter before it was
taken out of view and stopped being dealt with. If it wasn't this
way, you could rush through levels if the transition timer was high,
and then walk back to see different unfinished transitions for
each preious level, which would then transition again depending
on where you are standing.

It would be good to be able to handle the unfinished transitions more
smoothly, ideally by queuing levels from last to current (so that any
overlaps are overwritten by the newer levels), and having them all
decrease their render counter over time. I've considered the option
of having them disappear when rendered off screen, but this would add
complications if one level was created sooner than a previous level,
and the newer level went off screen while it's previous level didn't
-- there would be a gap in the string of levels. So I think a queue,
would be better, as it can just drop them off as soon as their render
transition reaches zero.