Changed generate level to take a single pointer to the previous level (or NULL if none), and return a pointer to a newly allocated level, rather than taking a pointer to a previously existing level pointer and doing all of the allocations and setting data in the function.
This seems simpler, easier to debug and it just feels like there is less
to go wrong with this method. I did the opposite move with
calculate_next_offsets a few commits back, but after some consideration,
and listening to Casey Muratori's reasoning begind creating functions this
way -- making them as 'functional' as possible, less complex, etc -- I plan
to revert to the old way of just passing back an offset struct, rather than
doing all the manipulations and calculations in one function, like I had
been doing with generate_level and will aim to stick to this methodology
in future where possible.