Previous and next levels are now chained (a previous chain and a next chain) and levels are rendered in this order: - Last previous - Last previous -1, etc - Current level's previous level - Furthest next - Furthest next -1, etc - Current level's next level - Current level
It may be worth saving current level's previous level to be rendered 3rd to
last in the list, because it is a higher priority than 'furthest next', etc
(because it is more relevent to the player), but the current way does have
the advantage of a more forward facing 'flow', with the current level always
active -- it should always be rendered last (as to be on top of other pixels).

Following this logic, maybe it would be best to interweave next and previous
levels, i.e. rendering outer levels first (wherever prev_levels_to_render
is higher than next_levels_to_render, or vice versa, always go for the
highest number, defaulting to rendering prev_levels first, because --
as one must be prioritised -- looking forward will be generally more useful
than backwards). Then there would be a sort of 'level pyramid' in terms of
layers, outer levels below inner levels, before the current level's prev and
next, and then current level at the top.

Playing around with high transition times makes this quite interesting, but
the effect of the new levels overwriting older levels does look a bit
jarring... it just makes the level black and then slowly fades in to the
right colour. I'll have to play around with alpha values and see if there
is a way to smoothly fade over old tiles.