Added a new render depth counter for next and prev levels, which tracks the furthest level in the chain to have a render transition above 0.
This was added because in some cases where, say you are moving fowards and
traverse a large level, spend some time in it, and then quickly move past
a much smaller level. If the larger level had a higher render transition
than the smaller one after the player moved on beyond both levels, the
larger level would not complete it's transition and would disappear
suddenly as the smaller level completed it's transition to zero vis.
Upon backtracking, triggering the larger level would cause it's transition
to continue from where it left off, as it never had a chance to complete.

Now, 'next_render_depth' and 'prev_render_depth' will hold the index of
that larger level until its transition completes. The program will loop
through the chain of levels looking for the furthest level with a
render transition above zero, but only as far as the render depth,
and then update it depending on how deep in the loop it found the last
level with a render transition until that point. This means the whole
chain of levels doesn't need to be looped through each time in order
to determine which levels to render -- only the active levels, or levels
active last frame will be looped through to determine 'render depth'.