Now handling pixel-by-pixel movement, with transitions to next and previous levels as before. This happens by structuring player position by tile_x/y and pixel_x/y. A player still belongs to tile x and y, however, a pixel x and y value is now stored, which relates to an offset within the tile.
This makes it incredibly simple to retain correct positioning between
tiles, and the conditions for detecting when to stop at a wall, when to
know that a player is leaving an exit/entrance have all been simplified,
with the leverage of a few functions, all of which has broken down
the move_player code in to something much better flowing.

Collisions are currently only happening with the center of the player tile,
but this should be simple to correct.