Added linear alpha blending based on the level render gradient.
The create_colour function now takes a pointer to the current
position on the pixel buffer that the renderer is going to write
to, and calculates a linear blend depending on what has been
blitted previously.

Some of the combinations of colours don't look quite right, for
example when two walls are on top of each other, they should
display the same colour - but instead it seems slightly off, so
there is something to debug here. Aside from that I am pleased
with the initial results as it looks smoother than the
levels jutting out in front of one another.