Alpha blending is now working as it is supposed to. The bug was very obvious after a second look.
The equation for blending is:

r = new red colour
buffer_red = red colour in pixel buffer

result_red = buffer_red + (a * (r - buffer_red));

NOT

result_red = buffer_red + (a * (r - a));

The reference material I was referring to used 'a' to demark
'buffer red' as different colours were 'a' and 'b', vs 't' as
alpha scale.

(a * (r - buffer_red)) represents a scalar between new and old
colour, much like portion of distance might be calculated
('a' being percentage of distance between two objects, for example).
This is then added on to the old 'buufer' value, and alpha determines
the scale between the two values.