Flawed but working reflection vectors on wall collisions. Added vector2 struct and dot_product functions to deal with the calculations required for this.
There is still a bug when gliding over the wall of an exit, which
is ever more evident now, but this should go away once collisions
are of the whole shape of the player, rather than just the center
point -- which is where the detection for a new level happens.

The player will continue to move towards the wall as each iteration
of movement accelerates towards it, and then velocity is corrected
each time, which accounts for the slight stuttering when 'sliding'
against a wall. If I am not mistaken, this should be solvable
with a timestep-based collision detector, a la Handmade Hero,
which would move a player as far as the collision, change the
player delta to this new direction, and move them as far as the
remaining time step.