Now using a timestep based collision system, however there is a bug with getting stuck in a wall at the moment, which needs investigating (hence the debug printf's everywhere).
When hitting a wall, the nearest collision is detected, and a timestep
(t_min) returned, indicating the percentage of the proposed movement
to take, leaving a percentage behind for another step after wall
reflections (t_remaining). When t_remaining is 0, a full step
has been taken, and the iterations complete. It may be wise
to limit the number of iterations, as leaving limitless collision
reactions could prove problems when more complex collision
scenarios arise.