Fixed bug noted in last commit. This was due to the collision loop iterating over all tiles affected by 'new_player_position', delta x/y changing, and then the same tiles being checked against despite the fact that they didn't interfere on their perpendicular axis.
For example, if dx and dy both equal 1, and the player hits a bottom
tile, with the delta values projecting over a tile to the right of
the bottom tile it is touching, four tiles are checked against according
to the square this vector creates.

When the first loop is iterated over, y is reflected to 0, but x
remains at 1. The player is at the edge of the bottom tile it hit
at the nearest point the timestep allows. On the second iteration,
dx is 1*remaining_time, and should be free to move right, however
the buggy version of the code still checked the x coordinates against
the bottom right walls, and saw an intersection -- despite Y being
above it.

The simple fix was including a condition for the new proposed y position
being between the bounds of the top and bottom walls. If between the bounds,
it is a valid condition, otherwise it is not.

This fixed the issue, and the player seems to glide along the surface of
the walls without issue.