Added Minkowski addition collision detection for walls. Collision is queried against a single pixel point (player center), and wall radiuses are increased by half the player width / height, and collision is tested against those bounds.
It's working almost perfectly, and with the perpendicular wall
reflections and timestep correction the motion is nice and smooth,
however there is still a bug. I think it is something to do with the
out of bounds checking -- the last bug I fixed to enable this
functionality was because the collision detection algorithm
needs to know NOT to check out of bounds so the player can
get out of an exit -- however the vast majority of the time
there is no problem going in and out of exits.

Two manifestations of the bug I have seen so far are
a) Getting stuck on the tiles in between exit and entrances.
b) One occurance of getting stuck on a corner coming after coming
out of an exit. I am not sure if it is to with the exit, it seems
unlikely, but it might be worth bearing in mind.

I have also reduced the tile size to 10 pixels, and enabled
variable window sizes after setting the game resolution to
something low. As all graphics were blocks of pixels, this
should reduce the overhead substantially, and allow the
graphical side of the game to only produce as much overhead
as needed.

By default the SDL layer will determine a maximum pixel ratio
based on screen size and game resolution, and make a window of
the maximum size at correct resolution. The user can then increase
or decrease the window size by ratio with enter (+1) or backspace (-1).
Fullscreen is 'supported', however it will mean a skewed ratio for
the time being.