A number of small but important changes:
- Player position now rounds to the nearest integer on rendering,
rather than defaulting to floor (when (int)ing a float), meaning
that player now renders properly to the edge of a room, rather than
standing a pixel away because collision detection stops the player
from standing directly at a room edge.

- Corrected the player size to render it's actual size.

- Reduced the epsilon by 1dp. This SEEMS to have fixed to the bug
from last commit, where the player was getting stuck a lot --
especially on the right side of the player/map (left side of tiles).
Time will tell whether this was the entire problem, but the fact that
the left hand side of the tiles were sticking so much more than the right
with smaller epsilons is very suspect. Having an epsilon as small
as 0.00001 was fine for the right hand side of tiles / left of player,
but it takes 0.001 to perform correctly for the left side of tiles.
That's a whole 100* more/less accurate, which makes me uncomfortable.

I had a similar problem with my first attempt at new movement code,
where I was multiplying a friction constant by a floating acceleration/
velocity, and I got incredibly different results depending on whether
I was moving positive or negative on the axis. I believe I did something
like convert both to positive numbers, do the math, and then reconvert,
but I will need to investigate after this commit, as it could provide
a solution to the issue.

- Moved input processing outside of the pause loop.
When losing focus/pausing with keys down previously, letting go of keys,
and refocusing/unpausing, the player would career of in the previous
direction, for obvious reasons.