Fixed bug where collision code was not working correctly for odd player width/height due to integer rounding.
Also updated render round code, which rounds to the player pixel minus
t-epsilon*2, to account for the pixel integer rounding up and adding a
pixel of padding to the left/top of the player when an odd width/height.
When at a tile boundary, player is .5 distance away, meaning the pixel
rounds up to 1, adding padding/overflowing with integer rounding on
the render code.