Added collides property to entity and expanded collision detection to test for entities within level. Main collision detection algorithm now a seperate function which tests for entity collisions, so tilemap collisions are processed by creating a temporary entity, passing it the relevant information so it can be represented as an entity in terms of collision detection, and passed on to the collision detection function. Each entity in the current level is then looped through following this.
At the moment the system only seems to be working for centrally aligned
blocks (although it works for different sizes), so further testing needed.

Adding 0.002 to the entity render offset was also creating a jitter on the
movement, so I have removed it and movement is smooth again.
This was intended to offset the distance left behind by the collision
detection system when stopped against a wall -- for example being left
at position 5.49999 would round the player position on the renderer
down to 5. This was often proving problematic when adding changes to
the game, so I think I will leave this, as the small offset wasn't
really a proper solution.

There are a number of approaches to fixing this, two which come to mind
being a) a more sophisticated renderer which handles floating point
pixel positions, or b) making it so that players can go right up to
a wall position, somehow, and potentially make the collision checks
with rectangles 0.001 smaller than the actual size.