-Added basic (but still buggy) bullet projectile and expanded movement code to loop over player and bullet entities.
-Added a NULL entity to allow entity indexes of 0 to point to an entity that does nothing,
for uses such as parent entity which defaults to null, and then any index is an actual entity.

-Made a major bug fix in change_entity_location:
When taking entity index out of an old level block, function checked to see
if the block had run out of entities, and if so, took the next block as it's first block. This wasn't
checking to see if there was only a single block, however, meaning a segfault occured when a player
was taken out of a block in which he was the only entity. I'm not sure why this error hasn't occured
sooner, this should have failed every time the player went from level 0 to level 1 without any
additional entities in the block.

-Bullets currently reflect on the normal of the wall they collide with, while players still
reflect perpendicular to the normal relative to their velocity in that direction.

-create_entity has been simplified to take fewer arguments, allowing specific add_(entity_type) functions
to fill in specifics needed for the entity, leaving create_entity to be more general. This could use some
more work along these lines, I think.