Added level pointer to entity struct, and changed collision algorithm to test for entities in same level as target entity, rather than just the 'current level' (which I think will eventually be depreciated).
Entities now change levels using the same out of bounds/was on entrance/exit tile trigger
as player, but it merely updates their level pointer and the level block indexes, rather than
shifting the whole game to the new level.

As bullet is currently 2 pixels, there are some issues with the logic allowing out of bounds
movement at exits/entrances, causing bullets to occasionally get stuck in a wall block.
Two approaches for fixing this come to mind:
a) simply clamp the poxision of the entity to the bounds of the entrance/exit tile on level change.
b) make the collision algorithm also target next/prev level entities when on an entrance/exit tile,
rather than just testing for the level it is on.