Fixed bug where bullets were getting stuck in walls between levels. I think this had been around since some earlier builds, but having a wide player width helped stop it from cropping up as it would always collide with a wall on either side of the entrance/exit.
For smaller/faster entities, there was still a problem to be solved,
so now the entity collision algorithm detects whether the player
was at an entrance or exit, and going out of bounds, and if so,
tests an additional set of tiles in the new level, as if it was
coming in out of bounds.

This allows the walls on either side of an entrance/exit,
although technically seperated, to act as a continuous
collidable wall as far as the collision detection is concerned.

There does seem to be a segfault cropping up every now and again,
and I'm pretty sure it's only happened with the new out_of_bounds
looping in the collision detector, and usually when a load of bullets
firing.

Render order needs improving soon.
Bullets going out of the 'current level' are drawn over by the
drawing of the current level, which is later in the render loop.
As a start, it would be good to render level geometry, then entities.