Added visual health bar, and seperated movement code into function to allow cleaner conditions for entity movement specifics.
The seperation of move_entity / check_for_change_level is rather
ugly at the moment, but a better next/prev level solution has been
in order since the addition of entities, so it is a temporary placeholder
which just recreates the first coded solution for changing levels.

Once a generalised distance function has been added, this can
be collapsed into a single function and bullet_distance_sq
(which should be renamed) can be calculated independently.