Unrefined but working solution for collision testing entities in next level when on exit/prev level when on entrance. As with the rest of the move_entity function, this needs a lot of compressing down as there is way too much repetition of code and very specific conditionals.
One option I am considering is having two small arrays/lists for each
level which fill up with entities on/previously on entrance/exit tiles,
then have entrance and exit entities cross reference each other. It is
a tradeoff between having extra buffers to handle, which is slightly
fiddly, and passing in a whole level of entities wholesale, which is
a little over the top (something on the other side of the next map
doesn't need to be collision detected).