Basic vector field pathfinding for entities on same level as player.
Initially there was a bug with enemies disappearing when normalise
vector (after add_vectors) was included, and (I think) not without
this line – but after further testing I can't seem to recreate it,
so there could be a bug somewhere in the new code.

Next job is to enable vector fields for exits/entrances.

Currently, there is a vector of zero if a tile's left and right tiles
are equidistant from the player (x - x = 0, etc), and while not a big
problem as player movement or a broken block usually shifts the vector
field, it might be good to look at a couple of potential fixes: the best
I've looked at being dividing each tile by half on the vector map and doing
double the calculations.