Fixed pathfinding cases where velocity would zero because of equal distances on left and right of tile.
Algorithm now detects whether a tile to the right is less distance
than the tile being calculated (if velocity in x is zero), and gives
a positive velocity if true. The same for bottom/y.

I have also implemented a 'game pulse' which switches the position
on the player that is being calculated for v every 1/3 of a second.
This seems to help with navigating around corners, where velocity
values would conflict when adding velocities of different tiles.

There do seem to be cases where the enemies don't 'flock' very well,
especially when navigating between tight spaces, and get stuck trying
to move against each other, so a fix for that is something to think
about, although not the highest priority while the vast majority of
entities are moving correctly.

I've also thought aout implementing something to make groups of enemies
move less uniformly -- especially when a player moves. For example,
update distances of 1 on frame 1, distances of 2 on frame 2, etc,
so that enemies further out re-orient themselves to the new player
position slower than closer ones.