Implemented pathfinding for next/prev levels.
All enemy entities now move along their level's vector field, which
defaults to zero. When leaving a level, while a render offset is above
0.3 it will generate a vector field to the exit/entrance, meaning
players are triggered by a player and enemy coming into view.

The first few experiments with this made it obvious that this isn't a
particularly engaging mechanic, as there is a lot of stopping and
starting and it's very predictable. So to start off, whenever a level is
created there is a 50% chance that it will send enemies to the entrance.
This could either mean an instant onslaught of enemies, or double the
enemies in the next level. Obviously this is something to be worked on
heavily, but it seemed a worthwhile observation at the least.

There is still a problem with enemies getting permanenly stuck on each
other. I am unsure whether this is because of conflicting vectors
meaning they never move along the x axis, or because they have actually
found a way to overlap. The latter seems unlikely, but as previous bugs
have shown it could be possible. I haven't had a chance to debug the
values of each entity in a case like this, but took a short video to
document an instance of the problem:

https://streamable.com/j9tti

This and the last time this happened, both entities were crammed in a
single tile together by a collection of other entities, so it could be
something to do with the way enemies are flowing and aligning as a
result.

Having been able to see how the enemy pathfinding performs over multiple
levels like this, it seems like a rather flawed solution. Enemies kind
of bumble towards you awkwardly -- especially awkwardly around corners,
and vectors often point into walls, making the movement look slightly
unnatural.

This was a good first solution for getting basic pathfinding working,
but it will need a lot more work/thought/research to improve the
movement. One option might be some sort of flow/flocking control, so
that enemies don't all bundle into a tight space and barracade each
other into spaces of the map.

Another area that could do with work is my current debug workflow, which
is horribly inefficient at the moment. It would be good to start having
a proper interactive overlay to explore entitiy values, or even
something like live code editing to enable quicker resolution of
problems and a better debugging experience in the long term.