Reverted to simpler corner pathfinding handling.
I have done away with the use of 'game.pulse' for the time being, as I
am not liking the movement it gives entities around corners. Instead of
switching between three different points on an entity, I am averaging
their top_left / bottom_right points in the cases where x or y = 0.

This still strikes me as flawed, but it makes for smoother handling of
entities in general, but allows for more cases where entities get stuck
on corner tiles.

I plan to remedy this with a second layer of collision testing which
repels enemies from entities at a certain bounds, and after this will
perhaps integrate some sort of functionality to make enemies navigate
around each other to a destination. I aim to fix the corner cases with
the addition of these features, so using a pathfinding algorithm which
performs better for non-problematic cases seems like a better idea, as
future fixes will aim to smooth over the problematic cases, ideally
leading to smoother pathfinding all round.