Made walls real entities and refined move entity()
Before integrating more functionality into the move_entity()/collision
loops, it was obvious that the six(!) for loops needed refining as they
were problematic when it came to tweaking any of the intermediate
collision code (i.e. having to do it six times was a huge pain and
indicative of it being bad code).

After some consideration, it became obvious that creating 'fake wall
entities' on every loop of the collision code was quite strange logic,
and denied the reality that there just needed to be wall entities in the
game to match the way the code has evolved.

This is now a single loop, which is far more digestible -- the fact that
wall entities no longer need to be looped over individually makes it
much easier, as this is now part of a check for the next/prev levels
entities as well.

Move_entity() still has some conditional logic around the
collision/damage specifics that I am not fond of, but as with this
latest revision, this can be improved when a need for more complexity
asks for it.