Improved remove_entity workflow
This is the third (and should be final) fix for bugs which allowe
entities to be duplicated in game_state.entities. Firstly, there was an
erroneous conditional in the collision code where a bullet would lose
health instead of the entity it hit depending on which was first to be
registered by the collision loop. Fixing this alone should, I believe,
have fixed the case where some entities were being duplicated.

Any possibility of similar happenings in future is obviously bad, so I
have moved all conditions where an entity is removed to the end of the
main game loop, so an entity is only ever removed once and then never
again until it is a type other than entity_vacant.