Removed block_map and replaced with heat_map.
I am looking for a solution for handling pathfinding corner cases using
the heatmap to help an entity navigate by distance weights. This means
it needs permanent storage, which, like wall entities, isn't really a
problem because the values were being frequently created for calulation.

The main change here was changing the calculate_vector_field code to use
pointer arithmetic rather than array values, and blocks are now
registered on the main tile_map instead of in a seperate map field.
This was mainly to limit memory usage, and there is obviously the risk
that bugs in the block removal/addition code could trick the game into
thinking there is a wall where there isn't one, but this actually only
affects the vector field code now - very little relies on the tilemap,
so this may be something to consider doing away with in future.

An immediately positive result of this change is that the vector field
output code works anywhere, independent of a calculation.